# Async sampling