```
./target/profiling/mistralrs-bench -p 0 -g 64 -r 1 -c 8  gguf -t mistralai/Mistral-7B-Instruct-v0.1 -m TheBloke/Mistral-7B-Instruct-v0.1-GGUF -f mistral-7b-instruct-v0.1.Q4_K_M.gguf
```

Master

![image](https://github.com/EricLBuehler/mistral.rs/assets/12750442/dc6468fe-40f3-498d-ad53-feda993d4761)

This PR

![image](https://github.com/EricLBuehler/mistral.rs/assets/12750442/d72ad4f5-7dca-41d9-8789-58fffed9efd5)