@EricLBuehler 2024-04-26
> On bs=8 it improved from 5ms to 2ms.
Great! Perhaps we could profile it and see when the performance gains go away, and use this then.
> Clippy raised an issue of a non-async mutex being held across an await point.
We could use [this type](https://doc.servo.org/tokio/sync/struct.Mutex.html).

I think this sort of structure is very interesting, I'll take a look in the next few days. Thanks for working on it.