@lucasavila00 2024-04-26
I'll leave it open, so it picks up my commits.

I just fixed the clippy issue, by not holding the lock across awaits.