@lucasavila00 2024-04-26
I added 2 nvidia profiler screenshots to the main PR body.