@EricLBuehler 2024-04-26
@lucasavila00 

> cargo test fails because it can't download the mistral tokenizer from HF.

I set the `HF_TOKEN` secret during CI: 
> https://github.com/EricLBuehler/mistral.rs/blob/9ec90a40949efe9c68828c048da136874d1ba2f1/.github/workflows/ci.yml#L49