@lucasavila00 2024-04-26
> @lucasavila00, it looks like there are some merge conflicts.

I'm fixing it